universal-techland/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.6.0
firebase-admin==6.2.0
requests==2.31.0
feedparser==6.0.10
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.4
numpy==1.26.2
gunicorn==21.2.0
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.1.0
django-leaflet==0.30.0
